Updated Manager, started sortFaces in CubeCanvas
Added a 3D version of the game in manager
Added a sortFaces method in CubeCanvas and hooked it up to quicksort.
The sort does not work, it simply sorts it by index (where it starts off
sorted anyway). Need to make the method work off distance to the camera
and not index.