Pieces are now rendering
Added piece representations. Still need to add movement and texture
mapping. We have a potential problem that the current single
glDrawElements is using one opacity and not using textures at all. We
might have to break up the rendering every time we have to render a
piece for that. Will do research on textures.